+++ CHANGELOG.rst
@@ -35,3 +35,4 @@ Added
 - Optimizer function added.
 - Non-Max suppression (NMS) function added.
 - Mean average precision calculation added.
+- Loss function returns a dictionary of all losses.